pyproject.toml
[project]
name            = "prism"
version         = "0.1.0"
description     = "Volume periodicity analysis (Fourier + seasonality)"
requires-python = ">=3.8"
dependencies    = [
    "yfinance>=0.2",
    "alpaca-py>=0.10",
    "pandas>=2.0",
    "numpy>=1.23",
    "matplotlib>=3.8",
    "scipy>=1.11",
    "statsmodels>=0.14",
    "python-dotenv>=1.0"
]

[build-system]
requires      = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["prism*"]
exclude = ["data", "output"]
